integracaoTrelloAgente/agente04/requirements.txt

google-adk
py-trello
datetime
dotenv
